pip_requirements.txt
Django==1.11.5
django-model-utils==3.0.0
django-widget-tweaks==1.4.1
python-dateutil==2.6.1
pytz==2017.2

# Development
# ===================================================================
django-debug-toolbar==1.8
django-webtest==1.9.2
factory-boy==2.9.2
pytest==3.2.2
pytest-django==3.1.2
WebTest==2.0.28
